Split bundlereader into multiple files, organize funcs
